:dizzy:  feat: 외부 클릭 시 케밥 닫히게 수정
